LatestConfigs/twitter-api/requirements.txt
tweepy
kafka-python